backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.10
pydantic==2.5.0
python-multipart==0.0.6
reportlab==4.0.4